+++ R/zzz.R
@@ -9,14 +9,14 @@
   this.version = packageVersion("maotai")
   
   ## Print on Screen
-  packageStartupMessage("---------------------------------------------------------------------")
+  packageStartupMessage("**-----------------------------------------------------------------**")
   packageStartupMessage("** maotai")
   packageStartupMessage("**  - Tools for Matrix Algebra, Optimization and Inference Problems")
   packageStartupMessage("** Version    : ",this.version,"      (",this.year,")",sep="")
   packageStartupMessage("** Maintainer : Kisung You (kyoustat@gmail.com)")
   packageStartupMessage("**")
   packageStartupMessage("** Please share any bugs or suggestions to the maintainer.")
-  packageStartupMessage("---------------------------------------------------------------------")
+  packageStartupMessage("**-----------------------------------------------------------------**")
 }
 
 .onUnload <- function(libpath) {